+++ src/public.js
@@ -20,6 +20,7 @@ var Grid = function(element, options)
     this.current = 1;
     this.currentRows = [];
     this.identifier = null; // The first column ID that is marked as identifier
+    this.converter = null; // The converter for the column that is marked as identifier
     this.rowCount = ($.isArray(rowCount)) ? rowCount[0] : rowCount;
     this.rows = [];
     this.searchPhrase = "";
@@ -39,6 +40,21 @@ var Grid = function(element, options)
     // todo: implement cache
 };
 
+/**
+ * An object that represents the default settings.
+ * There are two ways to override the sub-properties.
+ * Either by doing it generally (global) or on initialization.
+ *
+ * @static
+ * @class defaults
+ * @for Grid
+ * @example
+ *   // Global approach
+ *   $.bootgrid.defaults.selection = true;
+ * @example
+ *   // Initialization approach
+ *   $("#bootgrid").bootgrid({ selection = true });
+ **/
 Grid.defaults = {
     navigation: 3, // it's a flag: 0 = none, 1 = top, 2 = bottom, 3 = both (top and bottom)
     padding: 2, // page padding (pagination)
@@ -67,18 +83,71 @@ Grid.defaults = {
      **/
     multiSelect: false,
 
+    /**
+     * Enables entire row click selection (`selection` must be set to `true` as well). Default value is `false`.
+     *
+     * @property rowSelect
+     * @type Boolean
+     * @default false
+     * @for defaults
+     * @since 1.1.0
+     **/
+    rowSelect: false,
+
     highlightRows: false, // highlights new rows (find the page of the first new row)
     sorting: true,
     multiSort: false,
     ajax: false, // todo: find a better name for this property to differentiate between client-side and server-side data
-    // post is obsolete (use instead requestHandler)
+
+    /**
+     * Enriches the request object with additional properties. Either a `PlainObject` or a `Function` 
+     * that returns a `PlainObject` can be passed. Default value is `{}`.
+     *
+     * @property post
+     * @type Object|Function
+     * @default function (request) { return request; }
+     * @for defaults
+     * @deprecated Use instead `requestHandler`
+     **/
     post: {}, // or use function () { return {}; } (reserved properties are "current", "rowCount", "sort" and "searchPhrase")
+
+    /**
+     * Sets the data URL to a data service (e.g. a REST service). Either a `String` or a `Function` 
+     * that returns a `String` can be passed. Default value is `""`.
+     *
+     * @property url
+     * @type String|Function
+     * @default ""
+     * @for defaults
+     **/
     url: "", // or use function () { return ""; }
+
     caseSensitive: true,
 
     // note: The following properties should not be used via data-api attributes
+
+    /**
+     * Transforms the JSON request object in what ever is needed on the server-side implementation.
+     *
+     * @property requestHandler
+     * @type Function
+     * @default function (request) { return request; }
+     * @for defaults
+     * @since 1.1.0
+     **/
     requestHandler: function (request) { return request; },
+
+    /**
+     * Transforms the response object into the expected JSON response object.
+     *
+     * @property responseHandler
+     * @type Function
+     * @default function (response) { return response; }
+     * @for defaults
+     * @since 1.1.0
+     **/
     responseHandler: function (response) { return response; },
+
     converters: {
         numeric: {
             from: function (value) { return +value; }, // converts from string to numeric
@@ -90,6 +159,14 @@ Grid.defaults = {
             to: function (value) { return value; }
         }
     },
+
+    /**
+     * Contains all css classes.
+     *
+     * @property css
+     * @type Object
+     * @for defaults
+     **/
     css: {
         actions: "actions btn-group", // must be a unique class name or constellation of class names within the header and footer
         center: "text-center",
@@ -112,16 +189,56 @@ Grid.defaults = {
         left: "text-left",
         pagination: "pagination", // must be a unique class name or constellation of class names within the header and footer
         paginationButton: "button", // must be a unique class name or constellation of class names within the pagination
+
+        /**
+         * CSS class to select the parent div which activates responsive mode.
+         *
+         * @property responsiveTable
+         * @type String
+         * @default "table-responsive"
+         * @for css
+         * @since 1.1.0
+         **/
         responsiveTable: "table-responsive",
+
         right: "text-right",
         search: "search form-group", // must be a unique class name or constellation of class names within the header and footer
-        selectCell: "select-cell", // must be a unique class name or constellation of class names within the entire table
-        selectBox: "select-box", // must be a unique class name or constellation of class names within the entire table
         searchField: "search-field form-control",
+        selectBox: "select-box", // must be a unique class name or constellation of class names within the entire table
+        selectCell: "select-cell", // must be a unique class name or constellation of class names within the entire table
+
+        /**
+         * CSS class to highlight selected rows.
+         *
+         * @property selected
+         * @type String
+         * @default "active"
+         * @for css
+         * @since 1.1.0
+         **/
+        selected: "active",
+
         sortable: "sortable",
         table: "bootgrid-table table"
     },
+
+    /**
+     * A dictionary of formatters.
+     *
+     * @property formatters
+     * @type Object
+     * @for defaults
+     * @since 1.0.0
+     **/
     formatters: {},
+
+    /**
+     * Contains all labels.
+     *
+     * @property labels
+     * @type Object
+     * @for defaults
+     **/
     labels: {
         all: "All",
         infos: "Showing {{ctx.start}} to {{ctx.end}} of {{ctx.total}} entries",
@@ -130,6 +247,14 @@ Grid.defaults = {
         refresh: "Refresh",
         search: "Search"
     },
+
+    /**
+     * Contains all templates.
+     *
+     * @property templates
+     * @type Object
+     * @for defaults
+     **/
     templates: {
         actionButton: "<button class=\"btn btn-default\" type=\"button\" title=\"{{ctx.text}}\">{{ctx.content}}</button>",
         actionDropDown: "<div class=\"{{css.dropDownMenu}}\"><button class=\"btn btn-default dropdown-toggle\" type=\"button\" data-toggle=\"dropdown\"><span class=\"{{css.dropDownMenuText}}\">{{ctx.content}}</span> <span class=\"caret\"></span></button><ul class=\"{{css.dropDownMenuItems}}\" role=\"menu\"></ul></div>",
@@ -148,7 +273,7 @@ Grid.defaults = {
         pagination: "<ul class=\"{{css.pagination}}\"></ul>",
         paginationItem: "<li class=\"{{ctx.css}}\"><a href=\"{{ctx.uri}}\" class=\"{{css.paginationButton}}\">{{ctx.text}}</a></li>",
         rawHeaderCell: "<th class=\"{{ctx.css}}\">{{ctx.content}}</th>", // Used for the multi select box
-        row: "<tr{{ctx.id}}>{{ctx.cells}}</tr>",
+        row: "<tr{{ctx.id}} class=\"{{ctx.css}}\">{{ctx.cells}}</tr>",
         search: "<div class=\"{{css.search}}\"><div class=\"input-group\"><span class=\"{{css.icon}} input-group-addon glyphicon-search\"></span> <input type=\"text\" class=\"{{css.searchField}}\" placeholder=\"{{lbl.search}}\" /></div></div>",
         select: "<input name=\"select\" type=\"{{ctx.type}}\" class=\"{{css.selectBox}}\" value=\"{{ctx.value}}\" />"
     }
@@ -366,8 +491,8 @@ Grid.prototype.select = function(rowIds)
 
             for (i = 0; i < this.selectedRows.length; i++)
             {
-                this.element.find("tbody > tr[data-row-id=\"" + this.selectedRows[i] + "\"] " + 
-                    selectBoxSelector).prop("checked", true);
+                this.element.find("tbody > tr[data-row-id=\"" + this.selectedRows[i] + "\"]")
+                    .addClass(this.options.css.selected).find(selectBoxSelector).prop("checked", true);
             }
 
             this.element.trigger("selected" + namespace, [selectedRows]);
@@ -419,8 +544,8 @@ Grid.prototype.deselect = function(rowIds)
             this.element.find("thead " + selectBoxSelector).prop("checked", false);
             for (i = 0; i < deselectedRows.length; i++)
             {
-                this.element.find("tbody > tr[data-row-id=\"" + deselectedRows[i][this.identifier] + "\"] " + 
-                    selectBoxSelector).prop("checked", false);
+                this.element.find("tbody > tr[data-row-id=\"" + deselectedRows[i][this.identifier] + "\"]")
+                    .removeClass(this.options.css.selected).find(selectBoxSelector).prop("checked", false);
             }
             
             this.element.trigger("deselected" + namespace, [deselectedRows]);